contract-generator/scripts/requirements.txt
python-docx>=0.8.11
pandas>=2.0.0
